// validator-our-ui/ui/src/common/types.ts
export interface ValidatorStats {
  index: number;
  publicKey: string;
  depositEvents: DepositEvent[];
  status?: string;
  balance: {
    eth: number | null; // 32.4523
    isExpected: boolean;
  };
}
// ...
export interface DepositEvent extends DepositEventArgs {
  blockNumber: number | undefined;
  transactionHash: string | undefined;
}

export interface NodeStats {
  peers: string[] | null;
  syncing: SyncStatus | null;
}

// Prysm deposit contract

export interface DepositEventArgs {
  pubkey: string; // '0xb01d89a3abf76b659e0ddfe7f08bc2df7900e70a9ac0dadef40ec4364cfc10bd679cf939b3497856f719101d33ef2eea',
  withdrawal_credentials: string; // "0x00b6589882996478845d4dd2ca85a57387d6a392217808c908add83b160a0fa7";
  amount: string; // "0x0040597307000000";
  signature: string; // "0x9085a737a4490a403e9d0773abcb283b39270a97df7e6fc95c10ac6e6ade3698a88d00b0712fd95b3c2c519035b829160efa34962c92d1dd440db532c5b9bdabf91c7927c3ca1350eb2eb0b52700abd2e704bb547a2dd1ecfa0368a4d72da5e6";
  index: string; // "0x6200000000000000";
}